+++ docs/changelog.rst
@@ -5,7 +5,7 @@ v0.8.0
 ^^^^^^
 
 * Dropped support for Python 3.6
-
+* Made ``Group.id`` auto-increment
 
 v0.7.4
 ^^^^^^